Address comment in PR #359
* Consolidate CLI tests for both ETL v1 and v2
